Migration/seed fungenti

+++ database/migrations/2021_06_11_094004_create_sponsorships_table.php
@@ -16,7 +16,7 @@ class CreateSponsorshipsTable extends Migration
         Schema::create('sponsorships', function (Blueprint $table) {
             $table->id();
             $table->string('name', 20)->nullable();
-            $table->tinyInteger('hours');
+            $table->smallInteger('hours');
             $table->float('amount', 5,2);
             $table->timestamps();
         });
